Move fsspecfsdp.py to demo/
